pyproject.toml
[tool.poetry]
name = "foodgram"
version = "0.1.0"
description = ""
authors = ["Aleksandr Menshikov <menshikov.a.v@icloud.com>"]


[tool.poetry.dependencies]
python = "^3.10"

Django = "3.2.5"
django-behaviors = "^0.5.1"
django-split-settings = "^1.1.0"
django-storages = "^1.12.3"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}
django-filter = "^22.1"
djangorestframework = "^3.13.1"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^1.7.0"}

environs = {extras = ["django"], version = "^9.5.0"}
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
model-bakery = "^1.6.0"
factory-boy = "^3.2.1"
djet = "^0.3.0"
Pillow = "^9.2.0"
pydantic = "^1.9.1"
python-slugify = "^6.1.2"
python-decouple = "^3.6"
structlog = "^21.5.0"
boto3 = "^1.24.39"
Collectfast = "^2.2.0"
djoser = "^2.1.0"
drf-extra-fields = "^3.4.0"
drf-writable-nested = "^0.6.4"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.5.0"
django-querycount = "^0.7.0"
django-coverage-plugin = "^2.0.3"
django-migration-linter = "^4.1.0"
django-extra-checks = "^0.11.0"
django-test-migrations = "^1.2.0"
nplusone = "^1.0.0"

pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
pytest-randomly = "^3.12.0"
pytest-timeout = "^2.1.0"
coverage = "^6.4.2"

mypy = "^0.961"
django-stubs = "^1.12.0"
types-python-slugify = "^6.1.0"

black = "^22.6.0"
autopep8 = "^1.6.0"
isort = "^5.10.1"

flake8 = "^4.0.1"
flake8-annotations = "^2.9.0"
flake8-annotations-coverage = "^0.0.6"
flake8-broken-line = "^0.4.0"
flake8-bugbear = "^22.7.1"
flake8-builtins = "^1.5.3"
pep8-naming = "^0.13.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-quotes = "^3.3.1"
flake8-todo = "^0.7"
flake8-isort = "^4.1.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
^/(
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | env
)/
  | config/settings
'''


[tool.autopep8]
max_line_length = 79
in-place = true
recursive = true
aggressive = 3
